requirements.txt
langdetect==1.0.7
python-telegram-bot==11.1.0
translate==3.5.0
